fixed some .env configuration by using port, also corrected status code 500

+++ backend/controller/controller.js
@@ -30,13 +30,16 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const { id } = req.params; // get the id from the request params
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(400).json({ success: false, message: "Invalid product id" });
+    }
     try {
         await Product.findByIdAndDelete(id); // find the product by id and delete it
         res.status(200).json({ success: true, message: "Product deleted successfully" });
     }
     catch (error) {
         console.error("Error in Deleting product:", error.message);
-        res.status(404).json({ success: false, message: "Product not found" });
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 }
